requirements.txt
Flask==2.0.2
flask_jwt_extended==4.3.1
flask_sqlalchemy==2.5.1
pytest==6.2.5
bcrypt==3.2.0
requests==2.26.0
psycopg2==2.9.1
redis==4.1.4
fastapi
uvicorn
aiopg
databases
fastapi_sqlalchemy
python-dotenv==0.19.2
aiohttp